+++ commitlint.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: ['@commitlint/config-conventional'],
+  rules: {
+    'signed-off-by': [2, 'always', 'Signed-off-by:'],
+  },
+};